ci: Add sccache to manylinux images
Adds sccache to our manylinux images, these are purposefully built
without the scccache-dist binary since we're not expecting to use that.

Another caveat of these builds is that they are built with the vendored
version of openssl.

This is to set the stage for us to be able to build binaries
sequentially.

Signed-off-by: Eli Uriegas <github@terriblecode.com>
ghstack-source-id: 59a0a2bcfa10d34d6978a3d38e00c90157a8c56c
Pull-Request: https://github.com/pytorch/pytorch/pull/156892
Signed-off-by: Eli Uriegas <eliuriegas@meta.com>